changed project structure and added jar with dependencies, also changed BusTicketService class structure.
