--- src/Identifiable.java
@@ -1,11 +1,0 @@
-public abstract class Identifiable {
-    protected String id;
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-}
