--- src/NullableWarning.java
@@ -1,6 +1,0 @@
-import java.lang.annotation.*;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.FIELD)
-public @interface NullableWarning {
-}
