--- src/Printable.java
@@ -1,3 +1,0 @@
-public interface Printable {
-    void print();
-}
