--- src/User.java
@@ -1,3 +1,0 @@
-public abstract class User {
-    public abstract void printRole();
-}
