+++ src/main/java/andersen/Identifiable.java
@@ -1,0 +1,13 @@
+package andersen;
+
+public abstract class Identifiable {
+    protected String id;
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+}
