+++ src/main/java/andersen/NullableWarning.java
@@ -1,0 +1,8 @@
+package andersen;
+
+import java.lang.annotation.*;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.FIELD)
+public @interface NullableWarning {
+}
