+++ src/main/java/andersen/Printable.java
@@ -1,0 +1,5 @@
+package andersen;
+
+public interface Printable {
+    void print();
+}
