+++ src/main/java/andersen/User.java
@@ -1,0 +1,5 @@
+package andersen;
+
+public abstract class User {
+    public abstract void printRole();
+}
